requirements.txt
confluent-kafka>=0.11.0
kafka-python>=1.4.4
pykafka>=2.6.0
colorlog
requests
